Rename osversion -> os_version
This module name is more readable
