+++ src/bci_build/package/apache_tomcat.py
@@ -3,8 +3,8 @@
 import datetime
 
 from bci_build.containercrate import ContainerCrate
-from bci_build.osversion import CAN_BE_LATEST_OS_VERSION
-from bci_build.osversion import OsVersion
+from bci_build.os_version import CAN_BE_LATEST_OS_VERSION
+from bci_build.os_version import OsVersion
 from bci_build.package import DOCKERFILE_RUN
 from bci_build.package import ApplicationStackContainer
 from bci_build.package import OsContainer
