+++ src/bci_build/package/base.py
@@ -5,7 +5,7 @@ from pathlib import Path
 
 from jinja2 import Template
 
-from bci_build.osversion import OsVersion
+from bci_build.os_version import OsVersion
 from bci_build.package import Arch
 from bci_build.package import BuildType
 from bci_build.package import OsContainer
