+++ src/bci_build/package/golang.py
@@ -4,8 +4,8 @@ import textwrap
 from itertools import product
 from typing import Literal
 
-from bci_build.osversion import CAN_BE_LATEST_OS_VERSION
-from bci_build.osversion import OsVersion
+from bci_build.os_version import CAN_BE_LATEST_OS_VERSION
+from bci_build.os_version import OsVersion
 from bci_build.package import DOCKERFILE_RUN
 from bci_build.package import LOG_CLEAN
 from bci_build.package import DevelopmentContainer
