+++ src/bci_build/package/kiwi.py
@@ -1,8 +1,8 @@
 """KIWI Appliances Builder SDK container for easy appliance building on SLE Micro."""
 
-from bci_build.osversion import ALL_NONBASE_OS_VERSIONS
-from bci_build.osversion import CAN_BE_LATEST_OS_VERSION
-from bci_build.osversion import OsVersion
+from bci_build.os_version import ALL_NONBASE_OS_VERSIONS
+from bci_build.os_version import CAN_BE_LATEST_OS_VERSION
+from bci_build.os_version import OsVersion
 from bci_build.package import BuildType
 from bci_build.package import DevelopmentContainer
 from bci_build.package import ParseVersion
