+++ src/bci_build/package/openjdk.py
@@ -4,7 +4,7 @@ import os
 from itertools import product
 from typing import Literal
 
-from bci_build.osversion import OsVersion
+from bci_build.os_version import OsVersion
 from bci_build.package import DOCKERFILE_RUN
 from bci_build.package import Arch
 from bci_build.package import DevelopmentContainer
