+++ src/bci_build/package/postgres.py
@@ -2,7 +2,7 @@
 
 from pathlib import Path
 
-from bci_build.osversion import OsVersion
+from bci_build.os_version import OsVersion
 from bci_build.package import DOCKERFILE_RUN
 from bci_build.package import ApplicationStackContainer
 from bci_build.package import ParseVersion
