+++ src/bci_build/package/python.py
@@ -4,9 +4,9 @@ import datetime
 from dataclasses import dataclass
 from typing import Literal
 
-from bci_build.osversion import CAN_BE_LATEST_OS_VERSION
-from bci_build.osversion import _SUPPORTED_UNTIL_SLE
-from bci_build.osversion import OsVersion
+from bci_build.os_version import CAN_BE_LATEST_OS_VERSION
+from bci_build.os_version import _SUPPORTED_UNTIL_SLE
+from bci_build.os_version import OsVersion
 from bci_build.package import DevelopmentContainer
 from bci_build.package import Replacement
 from bci_build.package import SupportLevel
