+++ src/bci_build/package/spack.py
@@ -1,8 +1,8 @@
 """HPC container for Spack package manager."""
 
-from bci_build.osversion import CAN_BE_LATEST_OS_VERSION
-from bci_build.osversion import _SUPPORTED_UNTIL_SLE
-from bci_build.osversion import OsVersion
+from bci_build.os_version import CAN_BE_LATEST_OS_VERSION
+from bci_build.os_version import _SUPPORTED_UNTIL_SLE
+from bci_build.os_version import OsVersion
 from bci_build.package import DOCKERFILE_RUN
 from bci_build.package import Arch
 from bci_build.package import DevelopmentContainer
