+++ src/dotnet/updater.py
@@ -22,8 +22,8 @@ from jinja2 import Template
 from packaging import version
 
 from bci_build.logger import LOGGER
-from bci_build.osversion import CAN_BE_LATEST_OS_VERSION
-from bci_build.osversion import OsVersion
+from bci_build.os_version import CAN_BE_LATEST_OS_VERSION
+from bci_build.os_version import OsVersion
 from bci_build.package import LOG_CLEAN
 from bci_build.package import DevelopmentContainer
 from bci_build.package import generate_disk_size_constraints
