+++ src/staging/bot.py
@@ -34,7 +34,7 @@ from obs_package_update.util import retry_async_run_cmd
 from oras.client import OrasClient
 
 from bci_build.logger import LOGGER
-from bci_build.osversion import OsVersion
+from bci_build.os_version import OsVersion
 from bci_build.package import ALL_CONTAINER_IMAGE_NAMES
 from bci_build.package import BaseContainerImage
 from dotnet.updater import DOTNET_CONTAINERS
@@ -1560,7 +1560,7 @@ def main() -> None:
     import sys
     from typing import Any
 
-    from bci_build.osversion import ALL_OS_VERSIONS
+    from bci_build.os_version import ALL_OS_VERSIONS
     from staging.build_result import is_build_failed
     from staging.build_result import render_as_markdown
