+++ tests/conftest.py
@@ -9,7 +9,7 @@ import pytest
 from _pytest.fixtures import SubRequest
 from _pytest.python import Metafunc
 
-from bci_build.osversion import OsVersion
+from bci_build.os_version import OsVersion
 from bci_build.package import ApplicationStackContainer
 from bci_build.package import BaseContainerImage
 from bci_build.package import DevelopmentContainer
