+++ tests/test_bot.py
@@ -4,8 +4,8 @@ import pathlib
 import pytest
 import yaml
 
-from bci_build.osversion import ALL_NONBASE_OS_VERSIONS
-from bci_build.osversion import OsVersion
+from bci_build.os_version import ALL_NONBASE_OS_VERSIONS
+from bci_build.os_version import OsVersion
 from staging.bot import StagingBot
