+++ tests/test_build_recipe.py
@@ -2,7 +2,7 @@ from datetime import date
 
 import pytest
 
-from bci_build.osversion import OsVersion
+from bci_build.os_version import OsVersion
 from bci_build.package import Arch
 from bci_build.package import BuildType
 from bci_build.package import DevelopmentContainer
