+++ tests/test_service.py
@@ -1,5 +1,5 @@
 from bci_build.containercrate import ContainerCrate
-from bci_build.osversion import OsVersion
+from bci_build.os_version import OsVersion
 from bci_build.package import BuildType
 from bci_build.package import DevelopmentContainer
 from bci_build.package import ParseVersion
